test: replace the hardcoded EV_MAX array with the libevdev one
Kernel 5.8 changed the value of SW_MAX but libevdev has a built-in check for
out-of-range bits (it silently discards those).

So where evemu is built against a new kernel but a libevdev that hasn't been
updated yet, our test expects bit up to the new SW_MAX while libevdev
discards all bits above the old SW_MAX.

Fix this by making sure the test and uses the same value of the
respective type max as evemu itself does.

Fixes #3

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>